@@ -10,26 +10,26 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@prisma/client": "^4.16.1",
-    "@types/node": "20.3.2",
+    "@prisma/client": "^5.0.0",
+    "@types/node": "20.4.1",
     "@types/react": "18.2.14",
     "@types/react-dom": "18.2.6",
     "@vercel/analytics": "^1.0.1",
     "axios": "^1.4.0",
-    "eslint": "8.43.0",
-    "eslint-config-next": "13.4.7",
-    "next": "13.4.7",
+    "eslint": "8.44.0",
+    "eslint-config-next": "13.4.9",
+    "next": "13.4.9",
     "next-pwa": "^5.6.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "sass": "^1.63.6",
-    "sharp": "^0.32.1",
+    "sharp": "^0.32.2",
     "the-new-css-reset": "^1.9.0",
-    "typescript": "5.1.5",
+    "typescript": "5.1.6",
     "xml2js": "^0.6.0"
   },
   "devDependencies": {
     "@types/xml2js": "^0.4.11",
-    "prisma": "^4.16.1"
+    "prisma": "^5.0.0"
   }
 }
